Refactor: 등록/수정 api 파일 정리
activityForm.ts 파일에 모두 모아 놓음
