--- app/_apis/activities/patchActivity.ts
@@ -1,25 +1,0 @@
-import axios from 'axios';
-
-import type { ActivityEdit } from '@/_types/activities/form.types';
-
-import axiosInstance from '@/_libs/axios';
-
-interface PatchActivityProps {
-  body: ActivityEdit;
-  id: string;
-}
-
-const patchActivity = async ({ id, body }: PatchActivityProps) => {
-  try {
-    const result = await axiosInstance.patch(`/my-activities/${id}`, body);
-    return result;
-  } catch (error) {
-    if (axios.isAxiosError(error)) {
-      throw error.response;
-    } else {
-      throw error;
-    }
-  }
-};
-
-export default patchActivity;
